Split up BitbucketBambooUpdateService and BambooBuildPlanUpdateProvider
